added functionality for receiving and saving messages

+++ src/main/java/com/example/b9connect/repos/FaqsRepository.java
@@ -11,5 +11,5 @@ import java.util.Set;
 @Repository
 public interface FaqsRepository extends JpaRepository<Faqs, Long> {
     @Query("SELECT f from Faqs f where f.service.id = :id")
-    Set<Faqs> findAllByServiceId(Integer id);
+    Set<Faqs> findAllByServiceId(Long id);
 }
